next item predictor module moved out of search namespace
