+++ pyproject.toml
@@ -50,7 +50,8 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
 python_search = 'python_search.cli:main'
-next_item_pipeline = 'python_search.search.next_item_predictor.next_item_pipeline:main'
+next_item_pipeline = 'python_search.next_item_predictor.next_item_pipeline:main'
+next_item = 'python_search.next_item_predictor.next_item_pipeline:main'
 collect_input = 'python_search.apps.collect_input:main'
 clipboard = 'python_search.apps.clipboard:main'
 notify_send = 'python_search.apps.notification_ui:main'
