+++ python_search/data_ui/results_page.py
@@ -8,8 +8,8 @@ def load_results_page():
 
     keys = config.commands.keys()
 
-    from python_search.search.next_item_predictor.inference.inference import Inference
-    from python_search.search.next_item_predictor.inference.input import ModelInput
+    from python_search.next_item_predictor import Inference
+    from python_search.next_item_predictor import ModelInput
 
     st.write("### Prediction results")
     st.write("##### Production run: " + Inference.PRODUCTION_RUN_ID)
