+++ python_search/entry_description_generator/pipeline.py
@@ -5,7 +5,7 @@ import numpy as np
 from python_search.configuration.loader import ConfigurationLoader
 from keras import layers
 import keras
-from python_search.search.next_item_predictor.mlflow_logger import configure_mlflow
+from python_search.next_item_predictor import configure_mlflow
 
 
 def extract_string_from_entries(entries) -> List[str]:
