+++ python_search/next_item_predictor/features/entry_embeddings/__init__.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import copy
 
 from python_search.events.latest_used_entries import RecentKeys
-from python_search.search.next_item_predictor.features.entry_embeddings.entry_embeddings import (
+from python_search.next_item_predictor.features.entry_embeddings.entry_embeddings import (
     EmbeddingSerialization,
     RedisEmbeddingsReader,
     RedisEmbeddingsWriter,
