+++ python_search/next_item_predictor/inference/inference.py
@@ -7,7 +7,7 @@ from python_search.configuration.configuration import PythonSearchConfiguration
 from python_search.configuration.loader import ConfigurationLoader
 from python_search.infrastructure.performance import timeit
 from python_search.logger import setup_inference_logger
-from python_search.search.next_item_predictor.inference.input import ModelInput
+from python_search.next_item_predictor.inference.input import ModelInput
 
 logger = setup_inference_logger()
