+++ python_search/next_item_predictor/next_item_model_v2.py
@@ -6,10 +6,10 @@ from pyspark.sql import Window
 
 from python_search.configuration.loader import ConfigurationLoader
 from python_search.search.models import PythonSearchMLFlow
-from python_search.search.next_item_predictor.features.entry_embeddings import InferenceEmbeddingsLoader
-from python_search.search.next_item_predictor.features.entry_embeddings.entry_embeddings import \
+from python_search.next_item_predictor.features.entry_embeddings import InferenceEmbeddingsLoader
+from python_search.next_item_predictor.features.entry_embeddings.entry_embeddings import \
     create_key_indexed_embedding
-from python_search.search.next_item_predictor.model_interface import ModelInterface
+from python_search.next_item_predictor.model_interface import ModelInterface
 
 
 class NextItemModelV2(ModelInterface):
